requirements.txt
cryptography==44.0.2
pandas==1.2.3
pycryptodome==3.22.0
pytest==8.3.5
qiskit==0.25.0
qiskit_aer==0.8.0
qiskit_aqua==0.9.0
qiskit_ibmq_provider==0.12.2
qiskit_ignis==0.6.0
qiskit_terra==0.17.0
quantum_random==1.3.2
